+++ helper-scripts/src/deleteDevStackLegacy/deleteDevStacksLegacy.ts
