+++ helper-scripts/src/deleteDevStacks/getStacks/getStacks.ts
@@ -1,5 +1,6 @@
-import inquirer from "inquirer";
 import { $, chalk, echo } from "zx";
+import { PrioritisedStacks } from "../deleteDevStacks.js";
+import inquirer from "inquirer";
 
 export const protectedStacks = [
   "mob-sts-mock",
@@ -187,7 +188,7 @@ const checkIfProtectedStack = (stacks: string[], protectedStacks: string[]) => {
   });
 };
 
-const getStacks = async (): Promise<PrioritisedStacks> => {
+export const getStacks = async (): Promise<PrioritisedStacks> => {
   const selectedStacks: string[] = [];
 
   const baseStackName = await getBaseStackNames();
@@ -198,63 +199,3 @@ const getStacks = async (): Promise<PrioritisedStacks> => {
 
   return prioritiseStacks(selectedStacks);
 };
-
-const deleteStack = async (stackName: string): Promise<void> => {
-  try {
-    await $`./src/james_delete_stack.sh ${stackName}`;
-  } catch (error: unknown) {
-    echo(chalk.red(`Unable to delete stack: ${stackName}`));
-    echo(chalk.red(`Error: ${error}`));
-    process.exit(1);
-  }
-};
-
-const deleteStacks = async (stacks: PrioritisedStacks): Promise<void> => {
-  const {
-    stacksToDeleteOrder01,
-    stacksToDeleteOrder02,
-    stacksToDeleteOrder03,
-  } = stacks;
-
-  await Promise.all(
-    stacksToDeleteOrder01.map(async (stackName) => {
-      echo("");
-      echo(`Deleting stack: ${stackName}`);
-      await deleteStack(stackName);
-      echo(chalk.green.bold(`${stackName} stack deleted`));
-      echo("");
-    }),
-  );
-
-  await Promise.all(
-    stacksToDeleteOrder02.map(async (stackName) => {
-      echo(`Deleting stack: ${stackName}`);
-      await deleteStack(stackName);
-      echo(chalk.green.bold(`${stackName} stack deleted`));
-      echo("");
-    }),
-  );
-  await Promise.all(
-    stacksToDeleteOrder03.map(async (stackName) => {
-      echo(`Deleting stack: ${stackName}`);
-      await deleteStack(stackName);
-      echo(chalk.green.bold(`${stackName} stack deleted`));
-      echo("");
-    }),
-  );
-};
-
-export const deleteDevStacks = async (): Promise<void> => {
-  try {
-    const stacks = await getStacks();
-    await deleteStacks(stacks);
-  } catch (error: unknown) {
-    echo(chalk.red("There was an error. Error:", error));
-  }
-};
-
-interface PrioritisedStacks {
-  stacksToDeleteOrder01: string[];
-  stacksToDeleteOrder02: string[];
-  stacksToDeleteOrder03: string[];
-}
